Refactor code to get first & last faster
Instead of extracting all numbers out of row,
refactored to find the first number.
If one is found then stop searching.
Do the same but look for the last number by parsing characters from
the end of the row backwards.
This means that when a row contains more than 2 numbers
that the characters between the 2 don't need to be parsed.
Tests updated.

Also fixed bug where the DisplayLog toggle was only being applied to INFO logs,
when these are the ones that are only useful when debugging.